Small Changes.
Uploading to not lose track of whats been updated.

+++ malugri-mobile/librevolution/brstm-main.cpp
@@ -118,10 +118,10 @@ extern "C" unsigned long gHEAD1_blocks_samples(){
 
 
 extern "C" int16_t**  getBufferBlock(unsigned long sampleOffset){
-    unsigned long readLength;
+    unsigned long readLength; // simple fix debugger attatcher.
     if (sampleOffset/brstmp->blocks_samples < (brstmp->total_blocks)) readLength = brstmp->blocks_samples;
     else readLength = brstmp->final_block_size;
-    brstm_fstream_getbuffer(brstmp, brstmfile, sampleOffset, readLength);
+    brstm_fstream_getbuffer(brstmp, brstmfile, sampleOffset, static_cast<unsigned int>(readLength));
     return brstmp->PCM_buffer;
 }
 
@@ -173,4 +173,5 @@ DEBUGGING INSTRUCTIONS:
 
 4. Enable Address Sanitizer in your project scheme for advanced runtime diagnostics. It can help catch heap and stack memory errors.
 */
+
